Fixing bugs
