+++ Detectors/FIT/FT0/calibration/CMakeLists.txt
@@ -36,6 +36,7 @@ o2_add_library(FT0Calibration
       o2_add_executable(ft0-events-per-bc-processor
         COMPONENT_NAME calibration
         SOURCES workflow/FT0EventsPerBcProcessor-Workflow.cxx
+                src/EventsPerBcCalibrator.cxx
         PUBLIC_LINK_LIBRARIES
         O2::FT0Calibration
       )
